TUGAS/TUGAS_04/grpc_restAPI/grpc_implementation/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
locust
